REFACTOR test classes into `Unit` & `Feature` namespaces
